Função get_ini_configs movida para o módulo de arquivos.ini
