--- settings/arduino.ini
@@ -1,15 +1,0 @@
-[INPUTS]
-sensor_1 = ==|1
-sensor_2 = off
-sensor_3 = off
-digital_1 = off
-digital_2 = off
-
-[OUTPUTS]
-inverter = on
-lights = on
-buzzer = on
-valve_1 = on
-valve_2 = off
-emergency = off
-
