--- settings/configs.ini
@@ -1,7 +1,0 @@
-[GERAIS]
-nome = FONTES/TELEFONES
-parada = 1.87
-
-
-
-
