--- settings/users.json
@@ -1,10 +1,0 @@
-{
-   "users": [
-      {
-         "name": "LUCAS DA SILVEIRA",
-         "id": "208417097",
-         "editor": true,
-         "operator": true
-      }
-   ]
-}
